move old to legacy
